refactor(slack): extract socket reconnect policy helpers
(cherry picked from commit 1dd77e410658efb23b78b30d8e5e2ae03f71a8d5)
